Add level #create feature spec
